Update installation instructions for Debian and Ubuntu
Very old versions of Debian and Ubuntu need a newer version of Git in
order to run. However, we no longer build packages for these versions
and the instructions are confusing to users. Remove the references to
installing newer versions of Git, since it's no longer necessary, and
since the instructions for Debian and Ubuntu are now the same, coalesce
them into one section.

Also, remove several instances of trailing whitespace to keep the file
tidy.